+++ prp/models/sample.py
@@ -1,5 +1,5 @@
 """Data model definition of input/ output data"""
-from typing import List, Union, Optional, Dict
+from typing import Dict, List, Optional, Union
 
 from pydantic import Field
 
@@ -35,7 +35,6 @@ class MethodIndex(RWModel):
 class SampleBase(RWModel):
     """Base datamodel for sample data structure"""
 
-    sample_id: str = Field(..., alias="sampleId", min_length=3, max_length=100)
     run_metadata: RunMetadata = Field(..., alias="runMetadata")
     qc: List[QcMethodIndex] = Field(...)
     species_prediction: SpeciesPredictionResult = Field(..., alias="speciesPrediction")
@@ -43,6 +42,7 @@ class SampleBase(RWModel):
 
 class ReferenceGenome(RWModel):
     """Reference genome."""
+
     name: str
     accession: str
     fasta: str