+++ CHANGELOG.md
@@ -3,11 +3,14 @@
 ### Added
 
  - Added `sequencing_run` and `lims_id` to output (bonsai input)
+ - Added function for annotating delly variants intersecting with resistance targets
  - Added capability of handling empty dictionaries from serotypefinder output
 
 ### Fixed
 
 ### Changed
+
+ - Removed sample id from Sample object
 
 ## [0.6.0]